test: move helloworld to integration test
